+++ src/main/java/com/example/splitwise/service/PercentSplitService.java
@@ -18,7 +18,7 @@ public class PercentSplitService extends SplitService {
     public Boolean validate() {
         int sum = 0;
 
-        for (Map.Entry<User, Integer> mapElement : expense.getShare().entrySet()) {
+        for (Map.Entry<User, Double> mapElement : expense.getShare().entrySet()) {
             sum += mapElement.getValue();
         }
 
@@ -29,7 +29,7 @@ public class PercentSplitService extends SplitService {
     public List<Balance> calculateSplit() {
         List<Balance> balances = new ArrayList<>();
 
-        for (Map.Entry<User, Integer> mapElement : expense.getShare().entrySet()) {
+        for (Map.Entry<User, Double> mapElement : expense.getShare().entrySet()) {
             balances.add(new Balance(expense.getPaidBy(), mapElement.getKey(),
                     (double) (mapElement.getValue() / 100) * expense.getAmount()));
         }
